+++ build_config.glibc-2.14.rb
@@ -85,6 +85,7 @@ MRuby::CrossBuild.new('x86_64-w64-mingw32') do |conf|
 
   [conf.cc, conf.linker].each do |cc|
     cc.command = 'x86_64-w64-mingw32-gcc'
+    cc.flags << '-DPCRE_STATIC'
   end
   conf.cxx.command      = 'x86_64-w64-mingw32-cpp'
   conf.archiver.command = 'x86_64-w64-mingw32-gcc-ar'
@@ -101,6 +102,7 @@ MRuby::CrossBuild.new('i686-w64-mingw32') do |conf|
 
   [conf.cc, conf.linker].each do |cc|
     cc.command = 'i686-w64-mingw32-gcc'
+    cc.flags << '-DPCRE_STATIC'
   end
   conf.cxx.command      = 'i686-w64-mingw32-cpp'
   conf.archiver.command = 'i686-w64-mingw32-gcc-ar'